+++ packages/atomic/src/components/ipx/atomic-ipx-refine-modal/atomic-ipx-refine-modal.tsx
@@ -1,0 +1,170 @@
+import {
+  BreadcrumbManager,
+  BreadcrumbManagerState,
+  buildBreadcrumbManager,
+  buildFacetManager,
+  buildQuerySummary,
+  FacetManager,
+  QuerySummary,
+  QuerySummaryState,
+} from '@coveo/headless';
+import {Component, h, State, Prop, Element, Watch, Host} from '@stencil/core';
+import {rectEquals} from '../../../utils/dom-utils';
+import {
+  BindStateToController,
+  InitializableComponent,
+  InitializeBindings,
+} from '../../../utils/initialization-utils';
+import {Button} from '../../common/button';
+import {Hidden} from '../../common/hidden';
+import {
+  getClonedFacetElements,
+  RefineModalCommon,
+} from '../../common/refine-modal/refine-modal-common';
+import {Bindings} from '../../search/atomic-search-interface/atomic-search-interface';
+
+/**
+ * @internal
+ */
+@Component({
+  tag: 'atomic-ipx-refine-modal',
+  styleUrl: 'atomic-ipx-refine-modal.pcss',
+  shadow: true,
+})
+export class AtomicIPXRefineModal implements InitializableComponent {
+  @InitializeBindings() public bindings!: Bindings;
+  @Element() public host!: HTMLElement;
+
+  @BindStateToController('querySummary')
+  @State()
+  public querySummaryState!: QuerySummaryState;
+
+  @BindStateToController('breadcrumbManager')
+  @State()
+  public breadcrumbManagerState!: BreadcrumbManagerState;
+
+  @State()
+  public error!: Error;
+
+  @State()
+  public interfaceDimensions?: DOMRect;
+
+  @Prop({mutable: true}) openButton?: HTMLElement;
+
+  @Prop({reflect: true, mutable: true}) isOpen = false;
+
+  private breadcrumbManager!: BreadcrumbManager;
+  public querySummary!: QuerySummary;
+  private facetManager!: FacetManager;
+
+  @Watch('isOpen')
+  watchEnabled(isOpen: boolean) {
+    if (isOpen) {
+      if (!this.host.querySelector('div[slot="facets"]')) {
+        this.host.append(
+          getClonedFacetElements(
+            this.bindings.store.getFacetElements(),
+            this.facetManager
+          )
+        );
+      }
+      this.onAnimationFrame();
+    }
+  }
+
+  private onAnimationFrame() {
+    if (!this.isOpen) {
+      return;
+    }
+    if (this.dimensionChanged()) {
+      this.updateDimensions();
+    }
+    window.requestAnimationFrame(() => this.onAnimationFrame());
+  }
+
+  private dimensionChanged() {
+    if (!this.interfaceDimensions) {
+      return true;
+    }
+
+    return !rectEquals(
+      this.interfaceDimensions,
+      this.bindings.interfaceElement.getBoundingClientRect()
+    );
+  }
+
+  public updateDimensions() {
+    this.interfaceDimensions = this.bindings.interfaceElement
+      .getElementsByTagName('atomic-ipx-layout')[0]
+      .getBoundingClientRect();
+  }
+
+  public initialize() {
+    this.breadcrumbManager = buildBreadcrumbManager(this.bindings.engine);
+    this.facetManager = buildFacetManager(this.bindings.engine);
+    this.querySummary = buildQuerySummary(this.bindings.engine);
+  }
+
+  private renderHeader() {
+    return (
+      <div class="w-full flex justify-between mb-3">
+        <h2 class="text-2xl font-bold truncate">
+          {this.bindings.i18n.t('filters')}
+        </h2>
+        {this.breadcrumbManagerState.hasBreadcrumbs && (
+          <Button
+            onClick={() => this.breadcrumbManager.deselectAll()}
+            style="text-primary"
+            text={this.bindings.i18n.t('clear-all-filters')}
+            class="px-2 py-1"
+          ></Button>
+        )}
+      </div>
+    );
+  }
+
+  private renderBody() {
+    if (!this.bindings.store.getFacetElements().length) {
+      return <Hidden></Hidden>;
+    }
+
+    return (
+      <aside slot="body" class="flex flex-col w-full adjust-for-scroll-bar">
+        {this.renderHeader()}
+        <slot name="facets"></slot>
+      </aside>
+    );
+  }
+
+  public render() {
+    return (
+      <Host>
+        {this.interfaceDimensions && (
+          <style>
+            {`atomic-modal::part(backdrop) {
+            top: ${this.interfaceDimensions.top}px;
+            left: ${this.interfaceDimensions.left}px;
+            width: ${this.interfaceDimensions.width}px;
+            height: ${this.interfaceDimensions.height}px;
+            }`}
+          </style>
+        )}
+        <RefineModalCommon
+          bindings={this.bindings}
+          host={this.host}
+          isOpen={this.isOpen}
+          onClose={() => (this.isOpen = false)}
+          querySummaryState={this.querySummaryState}
+          title={this.bindings.i18n.t('filters')}
+          openButton={this.openButton}
+        >
+          {this.renderBody()}
+        </RefineModalCommon>
+      </Host>
+    );
+  }
+
+  public componentDidLoad() {
+    this.host.style.display = '';
+  }
+}