+++ packages/atomic-react/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.27.16](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.27.15...@coveo/atomic-react@1.27.16) (2022-11-30)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.27.15](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.27.14...@coveo/atomic-react@1.27.15) (2022-11-28)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.27.14](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.27.13...@coveo/atomic-react@1.27.14) (2022-11-25)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.27.13](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.27.12...@coveo/atomic-react@1.27.13) (2022-11-24)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.27.12](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.27.11...@coveo/atomic-react@1.27.12) (2022-11-22)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
 ## [1.27.11](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.27.10...@coveo/atomic-react@1.27.11) (2022-11-15)
 
 **Note:** Version bump only for package @coveo/atomic-react